requirements.txt
# 사내 메신저 v4.1 필수 라이브러리
# 설치: pip install -r requirements.txt

# ============================================================================
# 필수 라이브러리
# ============================================================================
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio[client]>=5.11.0
simple-websocket>=0.10.0
httpx>=0.27.0
keyring>=25.0.0
pycryptodome>=3.18.0
PyQt6>=6.5.0
PySide6>=6.7.0
cryptography>=41.0.0
bcrypt>=4.0.0  # [v4.2] 비밀번호 해시 강화
Flask-Limiter>=3.3.0  # [v4.3] Rate Limiting
Flask-WTF>=1.1.0      # [v4.3] CSRF Protection
Flask-Compress>=1.13  # [v4.4] Response Gzip Compression
cachelib>=0.13.0      # [v4.6+] Flask-Session CacheLib backend
pytest>=7.4.0         # [v4.3] Unit Testing
pytest-flask>=1.2.0   # [v4.3] Flask Testing Comparison

# ============================================================================
# 고성능 동시 접속 지원 (선택, 권장)
# ============================================================================
# 수십~수백 명 동시 접속 시 필요
# GUI 모드에서는 자동 비활성화됨 (PyQt6 충돌 방지)
gevent>=23.0.0
gevent-websocket>=0.10.1

# ============================================================================
# 개발/디버깅용 (선택)
# ============================================================================
# pyinstaller>=5.13.0
# auto-py-to-exe>=2.40.0
